server/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
av==14.2.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
frozenlist==1.5.0
google-api-core==2.24.1
googleapis-common-protos==1.67.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
livekit==0.20.5
livekit-protocol==0.9.2
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
websockets==15.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
